Added comments to methods, adjusted WordCount, added Correlator
Added javadoc comments to code written by Ahn, and in the WordCount
class added some more helpful errors information when the commandline is
used. Created the Correlator class and working on it so it will use the
DocCorrelator class. Went through Hamlet and the new atlantis text files
and removed Gutenburg project information.