Memory Analysis Files, README.txt
Performed some memory analysis using the memory analyzer plugin. Had to
create dump file, so dump files and memory analysis files are in this
commit. Also, created README.txt file which is our analysis of the
project. Still writing the README.txt but should be done with the next
two or three hours.